+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 * Create `CapacityProviderService` to make it possible to use capacity provider strategies solving [#599](https://github.com/pulumi/pulumi-awsx/issues/599).
 * Ensure that `awsx.apigateway.APIArgs` `RestApiArgs` reflect the underlying pulumi-aws library
 * Upgrade to Go1.16
+* Add support for `GENEVE`, `UDP` and `TCP_UDP` in `awsx.lb.TargetGroup` Protocols
 
 ## 0.25.0 (2021-02-12)